pyproject.toml
[tool.poetry]
packages = [
  { include = "gendiff" },
]
name = "primbayev_gendiff"
version = "0.1.0"
description = ""
authors = ["Aldiyar Primbayev <primbayev.96aldi@gmail.com>"]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.dependencies]
python = "^3.6"
pytest = "^6.2.1"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
pytest-cov = "^2.10.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
